Moved the common folder into the newly created dflowfm folder and updated the references.
