+++ hydrolib/core/io/ini/util.py
@@ -8,7 +8,7 @@ from pydantic.class_validators import root_validator, validator
 from pydantic.fields import ModelField
 from pydantic.main import BaseModel
 
-from hydrolib.core.io.common.models import LocationType
+from hydrolib.core.io.dflowfm.common.models import LocationType
 from hydrolib.core.utils import operator_str, str_is_empty_or_none, to_list
